requirements.txt
Django>=4.0.1,<4.1
celery==5.2.1
redis==4.5.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular==0.26.0
requests
django-cors-headers
